+++ saleor/graphql/page/tests/mutations/test_page_types_bulk_delete.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 import graphene
 import pytest
 
@@ -51,6 +53,45 @@ def test_page_type_bulk_delete_by_staff(
     assert data["count"] == page_type_count
 
     assert not Page.objects.filter(pk__in=pages_pks)
+
+
+@mock.patch("saleor.plugins.webhook.plugin.get_webhooks_for_event")
+@mock.patch("saleor.plugins.webhook.plugin.trigger_webhooks_async")
+def test_page_type_bulk_delete_trigger_webhooks(
+    mocked_webhook_trigger,
+    mocked_get_webhooks_for_event,
+    any_webhook,
+    staff_api_client,
+    page_type_list,
+    permission_manage_page_types_and_attributes,
+    settings,
+):
+    # given
+    mocked_get_webhooks_for_event.return_value = [any_webhook]
+    settings.PLUGINS = ["saleor.plugins.webhook.plugin.WebhookPlugin"]
+
+    staff_api_client.user.user_permissions.add(
+        permission_manage_page_types_and_attributes
+    )
+
+    page_type_count = len(page_type_list)
+    variables = {
+        "ids": [
+            graphene.Node.to_global_id("PageType", page_type.pk)
+            for page_type in page_type_list
+        ]
+    }
+
+    # when
+    response = staff_api_client.post_graphql(PAGE_TYPE_BULK_DELETE_MUTATION, variables)
+
+    # then
+    content = get_graphql_content(response)
+    data = content["data"]["pageTypeBulkDelete"]
+
+    assert not data["errors"]
+    assert data["count"] == page_type_count
+    mocked_webhook_trigger.call_count == page_type_count
 
 
 def test_page_type_bulk_delete_by_staff_no_perm(