+++ saleor/plugins/webhook/tests/subscription_webhooks/fixtures.py
@@ -479,6 +479,27 @@ def subscription_page_deleted_webhook(subscription_webhook):
 
 
 @pytest.fixture
+def subscription_page_type_created_webhook(subscription_webhook):
+    return subscription_webhook(
+        subscription_queries.PAGE_TYPE_CREATED, WebhookEventAsyncType.PAGE_TYPE_CREATED
+    )
+
+
+@pytest.fixture
+def subscription_page_type_updated_webhook(subscription_webhook):
+    return subscription_webhook(
+        subscription_queries.PAGE_TYPE_UPDATED, WebhookEventAsyncType.PAGE_TYPE_UPDATED
+    )
+
+
+@pytest.fixture
+def subscription_page_type_deleted_webhook(subscription_webhook):
+    return subscription_webhook(
+        subscription_queries.PAGE_TYPE_DELETED, WebhookEventAsyncType.PAGE_TYPE_DELETED
+    )
+
+
+@pytest.fixture
 def subscription_product_created_multiple_events_webhook(subscription_webhook):
     return subscription_webhook(
         subscription_queries.MULTIPLE_EVENTS,